fix: check for delete permission

+++ src/routes/(app)/(authenticated)/l/[listShortId]/delete/+page.server.ts
@@ -2,9 +2,14 @@ import type { PageServerLoad } from './$types';
 import { error, redirect } from '@sveltejs/kit';
 import { prisma } from '$lib/server/db';
 
-export const load: PageServerLoad = async ({ params }) => {
+export const load: PageServerLoad = async ({ params, parent }) => {
 	if (!params.listShortId) {
 		throw error(404);
+	}
+
+	const { permissions } = await parent();
+	if (!permissions.canDelete) {
+		throw error(403);
 	}
 
 	await prisma.list.delete({
